+++ CHANGES.rst
@@ -5,6 +5,9 @@ Changelog
 1.2.10 (unreleased)
 -------------------
 
+- Add upgrade step to install Kimug and migrate user to Keycloak.
+  [bsuttor]
+
 - WEB-4233 : Update GDPR text for new instances only
   [boulch]